backend/requirements.txt
fastapi
uvicorn[standard]
pyswisseph
timezonefinder
geopy
pydantic
httpx
python-dotenv
pytest
pytest-asyncio
